requirements.txt
line-bot-sdk
flask
openai < 1.0.0
gunicorn
mysql-connector-python
